Added Task category updation feature

+++ app/services/task_category_service.rb
@@ -6,11 +6,20 @@ class TaskCategoryService
   end
 
   def delete_all(name)
-    begin
-      task_category = TaskCategory.find_by(name: name)
+    task_category = TaskCategory.find_by(name: name)
+    if task_category
       task_category.destroy
-    rescue ActiveRecord::RecordNotFound => error
-      self.errors = error
+    else
+      self.errors = 'Taskcategory not Exist With this name'
+    end
+  end
+
+  def update(id, name, description)
+    task_category = TaskCategory.find_by(id: id)
+    if task_category
+      task_category.update(name: name, description: description)
+    else
+      self.errors = 'Taskcategory not exist with this id'
     end
   end
 end
